added tariffs archive sheet on tariffs list screen. Fixed TariffModelDto to TariffModel data conversion

+++ app/src/main/java/com/glazovnet/glazovnetapp/tariffs/data/entity/TariffModelDto.kt
@@ -12,7 +12,7 @@ data class TariffModelDto(
     val categoryCode: Int,
     val maxSpeed: Int,
     val costPerMonth: Int,
-    val prepaidTraffic: Int?,
+    val prepaidTraffic: Long?,
     val prepaidTrafficDescription: String?,
     val isForOrganization: Boolean
 )
